auth/requirements.txt
Flask>=2.1.3,<2.2.0
Flask-PyMongo>=2.3.0,<2.4.0
python-dotenv==0.20.0
webargs>=8.2.0,<8.3.0
marshmallow>=3.17.0,<3.18.0
pyjwt>=2.4.0,<2.5.0
requests>=2.28.1,<2.29.0
gunicorn==20.1.0
